@@ -8,39 +8,16 @@
  */
 package main.java.codedojo.kata.junittdd.diamond;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
 
-public class Main {
+class Main {
     private static final List<Character> ALPHABET =
-            new ArrayList<Character>(
-                    Arrays.asList(
-                            'A',
-                            'B',
-                            'C',
-                            'D',
-                            'E',
-                            'F',
-                            'G',
-                            'H',
-                            'I',
-                            'K',
-                            'L',
-                            'M',
-                            'N',
-                            'O',
-                            'P',
-                            'Q',
-                            'R',
-                            'S',
-                            'T',
-                            'V',
-                            'X',
-                            'Y',
-                            'Z'
-                    )
+            List.of(
+                    'A', 'B', 'C', 'D', 'E', 'F',
+                    'G', 'H', 'I', 'K', 'L', 'M',
+                    'N', 'O', 'P', 'Q', 'R', 'S',
+                    'T', 'V', 'X', 'Y', 'Z'
             );
 
     public static void main(String[] args) {
